separate oamhealth as a component for use in the footer
